lukas changes

+++ release.sh
@@ -24,8 +24,11 @@ fi
 echo $cmd_output
 cmd_output=$(/usr/bin/git push origin master 2>&1)
 echo $cmd_output
-cmd_output=$(/usr/bin/ssh -i /home/ubuntu/Acas-dev.pem ubuntu@34.243.107.7 'rm -f /var/www/html/.git/index.lock' 2>&1)
 cmd_output=$(/usr/bin/ssh -i /home/ubuntu/Acas-dev.pem ubuntu@34.243.107.7 'cd /var/www/html; git pull origin master' 2>&1)
+if [[ $cmd_output = *"index.lock': File exists"* ]]; then
+	cmd_output=$(/usr/bin/ssh -i /home/ubuntu/Acas-dev.pem ubuntu@34.243.107.7 'rm -f /var/www/html/.git/index.lock' 2>&1)
+	cmd_output=$(/usr/bin/ssh -i /home/ubuntu/Acas-dev.pem ubuntu@34.243.107.7 'cd /var/www/html; git pull origin master' 2>&1)
+fi
 if [[ $cmd_output = *"merge: composer.lock"* ]]; then
 	cmd_output=$(/usr/bin/ssh -i /home/ubuntu/Acas-dev.pem ubuntu@34.243.107.7 'rm -f /var/www/html/composer.lock' 2>&1)
 	cmd_output=$(/usr/bin/ssh -i /home/ubuntu/Acas-dev.pem ubuntu@34.243.107.7 'cd /var/www/html; git pull origin master' 2>&1)
@@ -35,3 +38,4 @@ echo -e "\nRunning 'composer update' and 'drush cr' on UAT. Please wait...\n"
 cmd_output=$(/usr/bin/ssh -i /home/ubuntu/Acas-dev.pem ubuntu@34.243.107.7 'cd /var/www/html; composer update; drush cr' 2>&1)
 echo $cmd_output
 echo -e "\nFinished release to UAT :)\n"
+
